refactor(lib) rewrite buildDownloadUrl
